+++ app/Models/Desa.php
@@ -11,4 +11,16 @@ class Desa extends Model
     use HasFactory;
 
     protected $table = 'desa';
+
+    // Specify which attributes are mass assignable
+    protected $fillable = [
+        'nama_desa',
+        'alamat_desa',
+        'kode_pos',
+        'telepon',
+        'email',
+    ];
+
+    // Optionally, specify the date fields for soft deletes
+    protected $dates = ['deleted_at'];
 }