+++ routes/web.php
@@ -13,13 +13,12 @@ use App\Http\Controllers\LoginSecurityController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AnggaranKasController;
 use App\Http\Controllers\ApbdesController;
-use App\Http\Controllers\DanasController;
-use App\Http\Controllers\DesasController;
-use App\Http\Controllers\DokumensController;
+use App\Http\Controllers\DanaController;
+use App\Http\Controllers\DesaController;
+use App\Http\Controllers\DokumenController;
 use App\Http\Controllers\NoRekeningsController;
 use App\Http\Controllers\RealisasiAnggaranController;
-use App\Http\Controllers\RealisasiAnggaransController;
-use App\Http\Controllers\Sp2dsController;
+use App\Http\Controllers\Sp2dController;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -154,15 +153,36 @@ Route::resource('anggaran-kas', AnggaranKasController::class);
 // APBDes Routes
 Route::resource('apbdes', ApbdesController::class);
 
-// Danas Routes
-Route::resource('danas', DanasController::class);
+// Dana Routes
+Route::middleware(['auth', 'XSS'])->group(function () {
+    Route::get('dana', [DanaController::class, 'index'])->name('danas.index');
+    Route::get('dana/create', [DanaController::class, 'create'])->name('danas.create');
+    Route::post('dana/store', [DanaController::class, 'store'])->name('danas.store');
+    Route::get('dana/{id}/edit', [DanaController::class, 'edit'])->name('danas.edit');
+    Route::put('dana/{id}', [DanaController::class, 'update'])->name('danas.update');
+    Route::delete('dana/{id}', [DanaController::class, 'destroy'])->name('danas.destroy');
+    Route::get('dana/{id}', [DanaController::class, 'show'])->name('danas.show');
+});
+
+// Route::resource(
+//     ['middleware' => ['auth', 'XSS']],
+//     function () {
+//         Route::get('dana', [DanaController::class, 'index'])->name('realisasi_anggarans.index');
+//         Route::get('dana/create', [DanaController::class, 'create'])->name('realisasi_anggarans.create');
+//         Route::post('dana/store', [DanaController::class, 'store'])->name('realisasi_anggarans.store');
+//         Route::get('dana/{id}/edit', [DanaController::class, 'edit'])->name('realisasi_anggarans.edit');
+//         Route::put('dana/{id}', [DanaController::class, 'update'])->name('realisasi_anggarans.update');
+//         Route::delete('dana/{id}', [DanaController::class, 'destroy'])->name('realisasi_anggarans.destroy');
+//         Route::get('dana/{id}', [DanaController::class, 'show'])->name('realisasi_anggarans.show');
+//     }
+// );
 
 // Desas Routes
-Route::put('desas/{desa}', [DesaController::class, 'update'])->name('desas.update');
+// Route::put('desas/{desa}', [DesaController::class, 'update'])->name('desas.update');
 
 
 // Dokumens Routes
-Route::resource('dokumens', DokumensController::class);
+Route::resource('dokumens', DokumenController::class);
 
 // No Rekenings Routes
 Route::resource('no_rekenings', NoRekeningsController::class);
@@ -171,20 +191,34 @@ Route::resource('no_rekenings', NoRekeningsController::class);
 // Route::resource('realisasi_anggarans', RealisasiAnggaransController::class);
 
 // SP2Ds Routes
-Route::resource('sp2ds', Sp2dsController::class);
+Route::resource('sp2ds', Sp2dController::class);
 
 // Route::group(
 //     ['middleware' => ['auth', 'XSS']],
 //     function () {
-//         Route::get('apbdes', [ApbdesController::class, 'index'])->name('apbdes.index');
-//         Route::get('apbdes/create', [ApbdesController::class, 'create'])->name('apbdes.create');
-//         Route::post('apbdes/store', [ApbdesController::class, 'store'])->name('apbdes.store');
-//         Route::get('apbdes/{id}/edit', [ApbdesController::class, 'edit'])->name('apbdes.edit');
-//         Route::put('apbdes/{id}', [ApbdesController::class, 'update'])->name('apbdes.update');
-//         Route::delete('apbdes/{id}', [ApbdesController::class, 'destroy'])->name('apbdes.destroy');
-//         Route::get('apbdes/{id}', [ApbdesController::class, 'show'])->name('apbdes.show');
+//         // Route::get('apbdes', [ApbdesController::class, 'index'])->name('apbdes.index');
+//         // Route::get('apbdes/create', [ApbdesController::class, 'create'])->name('apbdes.create');
+//         // Route::post('apbdes/store', [ApbdesController::class, 'store'])->name('apbdes.store');
+//         // Route::get('apbdes/{id}/edit', [ApbdesController::class, 'edit'])->name('apbdes.edit');
+//         // Route::put('apbdes/{id}', [ApbdesController::class, 'update'])->name('apbdes.update');
+//         // Route::delete('apbdes/{id}', [ApbdesController::class, 'destroy'])->name('apbdes.destroy');
+//         // Route::get('apbdes/{id}', [ApbdesController::class, 'show'])->name('apbdes.show');
+//         // Route::get('/apbdes/anggaran', [ApbdesController::class, 'showAnggaran'])->name('apbdes.anggaran');
+//         // Route::get('/apbdes/verifikasi', [ApbdesController::class, 'showVerifikasi'])->name('apbdes.verifikasi');
+//         // Route::get('/apbdes/realisasi', [ApbdesController::class, 'showRealisasi'])->name('apbdes.realisasi');
+//         // Route::post('/apbdes/verifikasi/toggle/{id}', [ApbdesController::class, 'toggleVerifikasi'])->name('apbdes.verifikasi.toggle');
+//         // Route::post('/apbdes/realisasi/toggle/{id}', [ApbdesController::class, 'toggleStatus'])->name('apbdes.status.toggle');
+
 //     }
 // );
+
+Route::group(['middleware' => ['auth', 'XSS']], function () {
+    Route::get('apbdes', [ApbdesController::class, 'index'])->name('apbdes.index');
+    Route::get('/apbdes/anggaran', [ApbdesController::class, 'showAnggaran'])->name('apbdes.anggaran');
+    Route::get('/apbdes/verifikasi', [ApbdesController::class, 'showVerifikasi'])->name('apbdes.verifikasi');
+    Route::get('/apbdes/realisasi', [ApbdesController::class, 'showRealisasi'])->name('apbdes.realisasi');
+});
+
 
 // Route::group(
 //     ['middleware' => ['auth', 'XSS']],
@@ -199,18 +233,18 @@ Route::resource('sp2ds', Sp2dsController::class);
 //     }
 // );
 
-// Route::group(
-//     ['middleware' => ['auth', 'XSS']],
-//     function () {
-//         Route::get('desas', [DesaController::class, 'index'])->name('desas.index');
-//         Route::get('desas/create', [DesaController::class, 'create'])->name('desas.create');
-//         Route::post('desas/store', [DesaController::class, 'store'])->name('desas.store');
-//         Route::get('desas/{id}/edit', [DesaController::class, 'edit'])->name('desas.edit');
-//         Route::put('desas/{id}', [DesaController::class, 'update'])->name('desas.update');
-//         Route::delete('desas/{id}', [DesaController::class, 'destroy'])->name('desas.destroy');
-//         Route::get('desas/{id}', [DesaController::class, 'show'])->name('desas.show');
-//     }
-// );
+Route::group(
+    ['middleware' => ['auth', 'XSS']],
+    function () {
+        Route::get('desas', [DesaController::class, 'index'])->name('desas.index');
+        Route::get('desas/create', [DesaController::class, 'create'])->name('desas.create');
+        Route::post('desas/store', [DesaController::class, 'store'])->name('desas.store');
+        Route::get('desas/{id}/edit', [DesaController::class, 'edit'])->name('desas.edit');
+        Route::put('desas/{id}', [DesaController::class, 'update'])->name('desas.update');
+        Route::delete('desas/{id}', [DesaController::class, 'destroy'])->name('desas.destroy');
+        Route::get('desas/{id}', [DesaController::class, 'show'])->name('desas.show');
+    }
+);
 
 // Route::group(
 //     ['middleware' => ['auth', 'XSS']],
@@ -238,57 +272,18 @@ Route::resource('sp2ds', Sp2dsController::class);
 //     }
 // );
 
-// Route::group(
-//     ['middleware' => ['auth', 'XSS']],
-//     function () {
-//         Route::get('sp2ds', [Sp2dController::class, 'index'])->name('sp2ds.index');
-//         Route::get('sp2ds/create', [Sp2dController::class, 'create'])->name('sp2ds.create');
-//         Route::post('sp2ds/store', [Sp2dController::class, 'store'])->name('sp2ds.store');
-//         Route::get('sp2ds/{id}/edit', [Sp2dController::class, 'edit'])->name('sp2ds.edit');
-//         Route::put('sp2ds/{id}', [Sp2dController::class, 'update'])->name('sp2ds.update');
-//         Route::delete('sp2ds/{id}', [Sp2dController::class, 'destroy'])->name('sp2ds.destroy');
-//         Route::get('sp2ds/{id}', [Sp2dController::class, 'show'])->name('sp2ds.show');
-//     }
-// );
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+Route::group(
+    ['middleware' => ['auth', 'XSS']],
+    function () {
+        Route::get('sp2ds', [Sp2dController::class, 'index'])->name('sp2ds.index');
+        Route::get('sp2ds/create', [Sp2dController::class, 'create'])->name('sp2ds.create');
+        Route::post('sp2ds/store', [Sp2dController::class, 'store'])->name('sp2ds.store');
+        Route::get('sp2ds/{id}/edit', [Sp2dController::class, 'edit'])->name('sp2ds.edit');
+        Route::put('sp2ds/{id}', [Sp2dController::class, 'update'])->name('sp2ds.update');
+        Route::delete('sp2ds/{id}', [Sp2dController::class, 'destroy'])->name('sp2ds.destroy');
+        Route::get('sp2ds/{id}', [Sp2dController::class, 'show'])->name('sp2ds.show');
+    }
+);
 Route::resource('tests', App\Http\Controllers\TestController::class)->middleware(['auth', 'XSS']);
 
 
@@ -647,7 +642,7 @@ Route::resource('desas', App\Http\Controllers\DesaController::class);
 Route::resource('dokumens', App\Http\Controllers\DokumenController::class);
 
 
-Route::resource('noRekenings', App\Http\Controllers\NoRekeningController::class);
+
 
 
 Route::resource('realisasiAnggarans', App\Http\Controllers\RealisasiAnggaranController::class);
@@ -657,3 +652,14 @@ Route::resource('sp2ds', App\Http\Controllers\Sp2dController::class);
 
 
 Route::resource('klasifikasiBelanjas', App\Http\Controllers\klasifikasi_belanjaController::class);
+
+
+
+
+Route::resource('kelompokNorekenings', App\Http\Controllers\kelompok_norekeningController::class);
+
+
+Route::resource('jenisNorekenings', App\Http\Controllers\jenis_norekeningController::class);
+
+
+Route::resource('detailNorekenings', App\Http\Controllers\detail_norekeningController::class);