+++ CHANGELOG.md
@@ -18,4 +18,43 @@
 -   Resolve conflict file ([49aa3d5389c96cf50d026897bc9b8b5de29c7ec9](https://github.com/MAGANG-SKSD/SKSD2/commit/49aa3d5389c96cf50d026897bc9b8b5de29c7ec9))
 
 -   Editing fitur Desa
+
+#### [(2024-09-17)]
+
+-   Editing function fitur desa
+
+#### [(2024-09-18)]
+
+-   Editing function fitur desa and realisasi anggaran ([538e0ebbc6a6968625e9fe9e6734d71827b12c92] https://github.com/MAGANG-SKSD/SKSD2/commit/538e0ebbc6a6968625e9fe9e6734d71827b12c92)
+
+-   Add function and moduls klasifikasi nomer rekening
+
+-   Add function in fitur dokumen
+
+-   Add index.blade.php view and create.blade.php in fitur APBDes
+
 -   Add UI apbdes ([cf1bb512d92a6b11dbc0edf1364bf771ba1053b1](https://github.com/MAGANG-SKSD/SKSD2/commit/cf1bb512d92a6b11dbc0edf1364bf771ba1053b1))
+
+-   Add UI apbdes ([cf1bb512d92a6b11dbc0edf1364bf771ba1053b1](https://github.com/MAGANG-SKSD/SKSD2/commit/cf1bb512d92a6b11dbc0edf1364bf771ba1053b1))
+
+#### [(2024-09-19)]
+
+-   Fix in web.php
+-   add model & controller Anggaran ([11db7b0129ad92b8af3a5bb4a1cf51bd1d993061](https://github.com/MAGANG-SKSD/SKSD2/commit/11db7b0129ad92b8af3a5bb4a1cf51bd1d993061))
+-   Perbaikan model dan penambahan controller kelompok ,jenis & Detail norekening ([11db7b0129ad92b8af3a5bb4a1cf51bd1d993061](11db7b0129ad92b8af3a5bb4a1cf51bd1d993061)))
+
+-   Fix it conflict modul, datatable, controller and view in APBDes
+
+-   Make view in dasboard and modified controller for dasboard
+
+-   Memperbaiki field tampilan view APBDes, menambahkan fungsi create dan memperbaiki fungsi edit
+
+#### [(2024-09-20)]
+
+-   Change database anggaran, anggaran model, apbdes model, and anggaran datatable. Addview for anggaran, verifikasi, and realisasi ([90c73473efdeae7fbb89456b915a21bb9d850458] https://github.com/MAGANG-SKSD/SKSD2/commit/90c73473efdeae7fbb89456b915a21bb9d850458)
+
+-   Changes database anggaran, model anggaran, controller anggaran ([b9cbdcbab2aa01c37b484e22f20de8806941b4d2] https://github.com/MAGANG-SKSD/SKSD2/commit/b9cbdcbab2aa01c37b484e22f20de8806941b4d2)
+
+-   Make view for APBDes in index and show blade ([10768bdd9dd7e4239851d82d0ffabf53e7834e98])
+
+-   Modified UI dashboard,controller and route